Break up SettingsNotice.
