+++ assets/js/components/SettingsNotice/index.js
@@ -1,0 +1,3 @@
+import SettingsNotice from './SettingsNotice';
+
+export default SettingsNotice;
