+++ assets/js/components/SettingsNotice/utils.js
@@ -1,0 +1,41 @@
+/**
+ * Settings notice utils.
+ *
+ * Site Kit by Google, Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Internal dependencies
+ */
+import InfoIcon from '../../svg/info-icon.svg';
+import SuggestionIcon from '../../svg/suggestion-icon.svg';
+import WarningIcon from '../../svg/warning-icon.svg';
+
+export const TYPE_WARNING = 'warning';
+export const TYPE_INFO = 'info';
+export const TYPE_SUGGESTION = 'suggestion';
+
+export const getIconFromType = ( type ) => {
+	switch ( type ) {
+		case TYPE_WARNING:
+			return <WarningIcon />;
+		case TYPE_INFO:
+			return <InfoIcon />;
+		case TYPE_SUGGESTION:
+			return <SuggestionIcon />;
+		default:
+			return null;
+	}
+};
